etl/requirements.txt
numpy
pandas
psycopg2-binary
ipython-sql
sql
notebook
sqlalchemy 
geopandas
matplotlib
folium